move sign_with_keys to top-level keys module
from crate::credit::keys
to crate::keys
